requirements.txt
asgiref==3.5.0
autopep8==1.6.0
certifi==2021.10.8
distlib==0.3.4
dj-database-url==0.5.0
Django==4.0.3
django-ckeditor==6.2.0
django-editorjs==0.2.1
django-froala-editor==4.0.10
django-js-asset==2.0.0
django-quill-editor==0.1.40
django-tinymce==3.4.0
djangorestframework==3.13.1
filelock==3.6.0
gunicorn==20.1.0
Jinja2==3.0.3
MarkupSafe==2.1.1
Pillow==9.0.1
pipenv==2022.1.8
platformdirs==2.5.1
psycopg2==2.9.3
pycodestyle==2.8.0
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tzdata==2021.5
Unidecode==1.3.4
virtualenv==20.13.3
virtualenv-clone==0.5.7
whitenoise==6.0.0
